Tag entities that have been upgraded to transient modifiers
Latest version of adorena uses transient modifiers instead of persistent
ones for scaling. For compatibility, old persistent modifiers were
removed on entity load or player join. However, this incurs a cost of
iterating over every attribute for each entity or player loaded.

This PR adds a PDC tag to entities that have already performed this
reset, and this reset is not ran again for that entity or player.
Note that the reset is run for every entity on config reload, since
the used attributes may have changed. It does not need to be run on
join or load, since the transient nature of these attributes ensures
that no stale ones are left behind.

Fixes #3